refactor: `WordTypes` を移動
