+++ test/user_dict/test_user_dict.py
@@ -5,9 +5,10 @@ from pathlib import Path
 import pytest
 from pyopenjtalk import g2p, unset_user_dict
 
-from voicevox_engine.model import UserDictWord, WordTypes
+from voicevox_engine.model import UserDictWord
 from voicevox_engine.user_dict.part_of_speech_data import (
     MAX_PRIORITY,
+    WordTypes,
     part_of_speech_data,
 )
 from voicevox_engine.user_dict.user_dict import (
